[FIX]  fix Merge conflict

+++ src/Render/Sprite.h
@@ -1,22 +1,23 @@
 #pragma once
 #include "Texture.h"
+#include "Core/include.h"
 #include <unordered_map>
 
 class Window;
 
 struct Animation
 {
-	int16 m_firstFrame = 0;
-	int16 m_lastFrame = 0;
+	int32 m_firstFrame = 0;
+	int32 m_lastFrame = 0;
 
-	int16 m_line = 0;
+	int32 m_line = 0;
 
-	int16 m_tileW = 0;
-	int16 m_tileH = 0;
+	int32 m_tileW = 0;
+	int32 m_tileH = 0;
 	
 	float32 m_duration = 0;
 
-	Animation(int16 firstFrame, int16 lastFrame, int16 line, int16 tileWidth, int16 tileHeight, float32 duration = 0.5f) :
+	Animation(int32 firstFrame, int32 lastFrame, int32 line, int32 tileWidth, int32 tileHeight, float32 duration = 0.5f) :
 		m_firstFrame(firstFrame),
 		m_lastFrame(lastFrame),
 		m_line(line),
@@ -31,14 +32,14 @@ class Sprite : public Texture
 {
 private:
 
-	int16 m_width = 0;
-	int16 m_height = 0;
+	int32 m_width = 0;
+	int32 m_height = 0;
 
 	std::unordered_map<std::string, Animation*> m_animationMap;
 
 	Animation* m_currentAnimation = nullptr;
-	int16 m_currentFrameX = 0;
-	int16 m_currentFrameY = 0;
+	int32 m_currentFrameX = 0;
+	int32 m_currentFrameY = 0;
 
 	float32 m_timer = 0.f;
